feat(component/favorites component/detail added beatiful alert in fav…

+++ client/src/components/User/Favorites/index.tsx
@@ -2,7 +2,10 @@ import { getFavManga, fetchDeleteFavorites } from '../../../features/user/userSl
 import { useEffect } from 'react';
 import { useAppDispatch, useAppSelector} from '../../../app/hooks';
 import useHeaders from "../../../app/headers";
-import { Link } from 'react-router-dom'
+import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
+import withReactContent from 'sweetalert2-react-content';
+import { RiErrorWarningFill } from 'react-icons/ri';
 
 
 
@@ -11,8 +14,55 @@ const Favorites = () =>{
     const { id, favorites, token} = useAppSelector(state=> state.user)
     const headers = useHeaders(token)
     const handleClick = (mangaid: any) => {
-        dispatch(fetchDeleteFavorites(id, mangaid, headers))
-        dispatch(getFavManga(id, headers))
+        const MySwal = withReactContent(Swal)
+    MySwal.fire({
+        html: <><RiErrorWarningFill size={55}/> <h1>Are you sure you want to delete this Manga fron favorites?</h1><h3>You won't be able to revert this!</h3></>,
+        showCloseButton: true,
+        focusConfirm: false,
+        showCancelButton: true,
+        background: "#212429",
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, cancel!',
+        confirmButtonAriaLabel: 'Delete it',
+        buttonsStyling: false,
+        customClass: {
+            confirmButton: 'confirmButtonDelete',
+            cancelButton: 'cancelButtonDelete'
+        }
+    }).then((result) => {
+        if (result.isConfirmed) {
+        if (mangaid === id) {
+            dispatch(fetchDeleteFavorites(id, mangaid, headers))
+            dispatch(getFavManga(id, headers))
+        }
+        MySwal.fire({
+            icon: 'success',
+            html: <><h1>Deleted!</h1><h3>This Manga has been deleted successfully.</h3></>,
+            focusConfirm: false,
+            confirmButtonColor: "#ea374b",
+            background: "#212429",
+            confirmButtonText: 'Ok',
+            buttonsStyling: false,
+            customClass: {
+            confirmButton: 'confirmButton',
+            }
+        })
+        } else if (
+            result.dismiss === Swal.DismissReason.cancel
+        ) {
+        MySwal.fire({
+            icon: 'error',
+            html: <><h1>Cancelled!</h1><h3>This Manga is safe :)</h3></>,
+            focusConfirm: false,
+            confirmButtonColor: "#ea374b",
+            background: "#212429",
+            confirmButtonText: 'Ok',buttonsStyling: false,
+            customClass: {
+            confirmButton: 'confirmButton',
+            }
+        })
+        }
+    })
     }
